fix: 修复 playground 运行报错的问题
