+++ src/frontend/ai-blueking/packages/ai-blueking/playground/views/CustomMessageSlotView.vue
@@ -92,6 +92,7 @@
   import { MessageRole, MessageStatus } from '@blueking/chat-helper';
 
   import CustomMessageRenderer from '../components/custom-widgets/CustomMessageRenderer.vue';
+  import { codeExample } from '../components/custom-widgets/custom-message-slot-example';
 
   import type { Message } from '@blueking/chat-x';
 
@@ -184,46 +185,6 @@
       }) as unknown as Message,
   );
 
-  const codeExample = `<template>
-  <ChatBot :url="apiUrl">
-    <!-- 使用 #message 插槽自定义消息渲染 -->
-    <template #message="{ message }">
-      <CustomMessageRenderer :message="message" />
-    </template>
-  </ChatBot>
-</template>
-
-<script setup>
-import { ChatBot } from '@blueking/ai-blueking';
-import CustomMessageRenderer from '../components/CustomMessageRenderer.vue';
-<\/script>
-
-<!-- CustomMessageRenderer.vue 核心逻辑 -->
-<script setup>
-import { computed } from 'vue';
-import { MessageRender } from '@blueking/chat-x';
-import { parseCustomBlocks } from './parse-custom-blocks';
-import ChartWidget from './ChartWidget.vue';
-import IframeWidget from '../components/IframeWidget.vue';
-import FormWidget from '../components/FormWidget.vue';
-
-const props = defineProps({ message: Object });
-const blocks = computed(() => parseCustomBlocks(props.message.content || ''));
-<\/script>
-
-<template>
-  <template v-for="(block, i) in blocks" :key="i">
-    <!-- 普通文本用 MessageRender 渲染 -->
-    <MessageRender
-      v-if="block.type === 'text'"
-      :message="{ ...message, content: block.content }"
-    />
-    <!-- 自定义组件分发渲染 -->
-    <ChartWidget v-else-if="block.data.type === 'chart'" :data="block.data" />
-    <IframeWidget v-else-if="block.data.type === 'iframe'" :data="block.data" />
-    <FormWidget v-else-if="block.data.type === 'form'" :data="block.data" />
-  </template>
-</template>`;
 </script>
 
 <style scoped>
